requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
gcloud==0.18.3
googleapis-common-protos==1.59.0
httplib2==0.21.0
idna==3.4
jws==0.1.3
numpy==1.21.6
oauth2client==4.1.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycryptodome==3.17
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==2.0.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
